+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - added add_path to wex::factory::link
 - added separate styling for not yet committed lines
 - added vim gJ command
+- used boost::parser
 
 ### Changed